% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyna-package.R
\docType{package}
\name{dyna-package}
\alias{dyna}
\alias{dyna-package}
\title{dyna}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
