% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyna-class.R
\name{DynaModel}
\alias{DynaModel}
\title{DynaModel R6 Class}
\description{
DynaModel R6 Class

DynaModel R6 Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DynaModel-new}{\code{DynaModel$new()}}
\item \href{#method-DynaModel-set_reader}{\code{DynaModel$set_reader()}}
\item \href{#method-DynaModel-set_parser}{\code{DynaModel$set_parser()}}
\item \href{#method-DynaModel-set_estimator}{\code{DynaModel$set_estimator()}}
\item \href{#method-DynaModel-read}{\code{DynaModel$read()}}
\item \href{#method-DynaModel-parse}{\code{DynaModel$parse()}}
\item \href{#method-DynaModel-estimate}{\code{DynaModel$estimate()}}
\item \href{#method-DynaModel-clone}{\code{DynaModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynaModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-DynaModel-new}{}}}
\subsection{Method \code{new()}}{
Create a new DynaModel object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynaModel$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The input \code{data} for the model.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynaModel-set_reader"></a>}}
\if{latex}{\out{\hypertarget{method-DynaModel-set_reader}{}}}
\subsection{Method \code{set_reader()}}{
Set the \code{parser} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynaModel$set_reader(fun, cols)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{A \code{function}.}

\item{\code{cols}}{TODO}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynaModel-set_parser"></a>}}
\if{latex}{\out{\hypertarget{method-DynaModel-set_parser}{}}}
\subsection{Method \code{set_parser()}}{
Set the \code{parser} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynaModel$set_parser(fun)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{A \code{function}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynaModel-set_estimator"></a>}}
\if{latex}{\out{\hypertarget{method-DynaModel-set_estimator}{}}}
\subsection{Method \code{set_estimator()}}{
Set the \code{estimator} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynaModel$set_estimator(fun, params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{A \code{function}.}

\item{\code{params}}{A \code{list} of additional estimator parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynaModel-read"></a>}}
\if{latex}{\out{\hypertarget{method-DynaModel-read}{}}}
\subsection{Method \code{read()}}{
Read raw data for processing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynaModel$read()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynaModel-parse"></a>}}
\if{latex}{\out{\hypertarget{method-DynaModel-parse}{}}}
\subsection{Method \code{parse()}}{
Parse data for estimation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynaModel$parse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynaModel-estimate"></a>}}
\if{latex}{\out{\hypertarget{method-DynaModel-estimate}{}}}
\subsection{Method \code{estimate()}}{
Estimate the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynaModel$estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynaModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DynaModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynaModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
