% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyna-class.R
\name{DynaNetwork}
\alias{DynaNetwork}
\title{DynaNetwork R6 Class}
\description{
DynaNetwork R6 Class

DynaNetwork R6 Class
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{n_nodes}}{Number of nodes in the network.}

\item{\code{n_edges}}{Number of edges in the network.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DynaNetwork-new}{\code{DynaNetwork$new()}}
\item \href{#method-DynaNetwork-set_data}{\code{DynaNetwork$set_data()}}
\item \href{#method-DynaNetwork-set_parser}{\code{DynaNetwork$set_parser()}}
\item \href{#method-DynaNetwork-set_estimator}{\code{DynaNetwork$set_estimator()}}
\item \href{#method-DynaNetwork-parse}{\code{DynaNetwork$parse()}}
\item \href{#method-DynaNetwork-estimate}{\code{DynaNetwork$estimate()}}
\item \href{#method-DynaNetwork-clone}{\code{DynaNetwork$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynaNetwork-new"></a>}}
\if{latex}{\out{\hypertarget{method-DynaNetwork-new}{}}}
\subsection{Method \code{new()}}{
Create a new DynaNetwork object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynaNetwork$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The input \code{data} for the model.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynaNetwork-set_data"></a>}}
\if{latex}{\out{\hypertarget{method-DynaNetwork-set_data}{}}}
\subsection{Method \code{set_data()}}{
Set the \code{data} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynaNetwork$set_data(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A \code{data.frame} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynaNetwork-set_parser"></a>}}
\if{latex}{\out{\hypertarget{method-DynaNetwork-set_parser}{}}}
\subsection{Method \code{set_parser()}}{
Set the \code{estimator} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynaNetwork$set_parser(fun)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{A \code{function}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynaNetwork-set_estimator"></a>}}
\if{latex}{\out{\hypertarget{method-DynaNetwork-set_estimator}{}}}
\subsection{Method \code{set_estimator()}}{
Set the \code{estimator} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynaNetwork$set_estimator(fun)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{A \code{function}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynaNetwork-parse"></a>}}
\if{latex}{\out{\hypertarget{method-DynaNetwork-parse}{}}}
\subsection{Method \code{parse()}}{
Parse \code{data} for estimation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynaNetwork$parse()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynaNetwork-estimate"></a>}}
\if{latex}{\out{\hypertarget{method-DynaNetwork-estimate}{}}}
\subsection{Method \code{estimate()}}{
Estimate the network model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynaNetwork$estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DynaNetwork-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DynaNetwork-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DynaNetwork$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
