% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyna-package.R
\docType{data}
\encoding{UTF-8}
\name{engagement}
\alias{engagement}
\title{Example Data on Student Engagement}
\format{
An \code{stslist} object (sequence data).
}
\source{
\doi{10.1016/j.compedu.2023.104787}
}
\usage{
engagement
}
\description{
Students' engagement states (Active / Average / Disengaged)
throughout a whole study program. The data was generated synthetically
based on the article "The longitudinal association between engagement and
achievement varies by time, students' profiles, and achievement state:
A full program study". Used also in the \code{tna} package.
}
\references{
Tikka S, López-Pernas S, Saqr M (2025).
"tna: An R Package for Transition Network Analysis."
\emph{Applied Psychological Measurement}. \doi{10.1177/01466216251348840}
}
\keyword{datasets}
