% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_model}
\alias{estimate_model}
\title{Estimate a Network Model}
\usage{
estimate_model(
  data,
  format = "sequences",
  type = "tna",
  cols = tidyselect::everything(),
  scaling = character(0L)
)
}
\arguments{
\item{data}{TODO}

\item{format}{TODO}

\item{type}{TODO}

\item{cols}{TODO}

\item{scaling}{TODO}
}
\value{
TODO
}
\description{
Estimate a Network Model
}
