% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyna-package.R
\docType{data}
\encoding{UTF-8}
\name{group_regulation}
\alias{group_regulation}
\title{Example Data on Group Regulation}
\format{
A \code{data.frame} object.
}
\source{
The data was generated synthetically.
}
\usage{
group_regulation
}
\description{
Students' regulation during collaborative learning. Students' interactions
were coded as: "adapt", "cohesion", "consensus", "coregulate", "discuss",
"emotion", "monitor", "plan", "synthesis". Used also in the \code{tna} package.
}
\references{
Tikka S, López-Pernas S, Saqr M (2025).
"tna: An R Package for Transition Network Analysis."
\emph{Applied Psychological Measurement}. \doi{10.1177/01466216251348840}
}
\keyword{datasets}
