% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_model}
\alias{estimate_model}
\title{Estimate a Network Model}
\usage{
estimate_model(data, cols, type = "transition_network")
}
\arguments{
\item{data}{TODO}

\item{cols}{TODO}

\item{type}{TODO}
}
\value{
TODO
}
\description{
Estimate a Network Model
}
